partial modules -- inexact and command tests failing, though
The nub of this change is that the current thread needs to know about
the currently running closure's frame and definition module (env) and
these must both be switched when the closure runs to be able to lookup
toplevel values in the right environment.  Previously we only handled
the frame.  That's it, one extra field!

The knock-on effect is over 4000 lines changed (about 8% of the code
base)!

As the code finally runs test.idio to completion (albeit with plenty
of errors) get a checkin done while we can!