job-control tests: fickle test subjects
In general, we're trying to provoke a specific error condition in a
pipeline or process substitution and see what Idio makes of it all
when it tries to read from said pipeline/named pipe.  However, things
do not go according to plan.

This is better but still misses a few transient cases.

Signed-off-by: Ian Fitchet <ian.fitchet@runscripts.com>