Woe is GC!
The previous commit prevented just under one million lexobj
struct-instances from being created which were never used (and each
struct-instance created four pairs in the call).

An unnoticed side-effect was that wobbled the GC into a bad place and
a normal test took 60% longer!  Fewer objects on the heap is not
necessarily a good thing, it seems.

It turns out that the GC is being called too often.  Case in point
being at the end of each "test-load" in lib/test.idio.

So this change is just some tweaks to when the GC is called and,
presumably, is just as liable to be wrong some time in the future.

Signed-off-by: Ian Fitchet <ian.fitchet@runscripts.com>