argument parsing: WIP
lib/SRFI-37.idio
lib/argparse.idio (a port of the Chicken Scheme args.scm)

At this point it is Scheme-oriented where arguments are passed as:

-dARG		and NOT -d ARG
--data=ARG	and NOT --data ARG

we want to rework the code in favour of the latter (a more familiar
form for Bash's getopts or Python's argparse).

And add some tests.

Signed-off-by: Ian Fitchet <ian.fitchet@runscripts.com>